+++ src/CSharp App/Program.cs
@@ -1,8 +1,6 @@
 ﻿﻿using Avalonia;
-using Serilog;
 using System;
-using System.IO;
-using VolumetricSelection2077.Converters;
+using VolumetricSelection2077.MessagePack.Helpers;
 
 namespace VolumetricSelection2077;
