+++ src/CSharp App/Services/GameFileService.cs
@@ -1,30 +1,17 @@
 using System;
 using System.Diagnostics;
 using System.IO;
-using System.Text.Json;
 using System.Threading.Tasks;
-using System.Xml.XPath;
-using MessagePack;
-using Microsoft.EntityFrameworkCore;
 using WolvenKit;
 using WolvenKit.Common.Services;
 using WolvenKit.Core.Interfaces;
 using WolvenKit.Core.Services;
 using WolvenKit.RED4.CR2W.Archive;
 using WolvenKit.RED4.CR2W;
-using SharpGLTF.Schema2;
-using VolumetricSelection2077.Helpers;
+using VolumetricSelection2077.MessagePack.Helpers;
 using VolumetricSelection2077.Models;
 using VolumetricSelection2077.Parsers;
 using WolvenKit.Common;
-using WolvenKit.Common.Conversion;
-using WolvenKit.Common.FNV1A;
-using WolvenKit.Common.PhysX;
-using WolvenKit.Modkit.RED4.Tools;
-using WolvenKit.RED4.Archive.CR2W;
-using WolvenKit.RED4.CR2W.JSON;
-using WolvenKit.RED4.Types;
-using JsonSerializer = System.Text.Json.JsonSerializer;
 
 namespace VolumetricSelection2077.Services;
